pyproject.toml
[tool.poetry]
name = "sibdev-job"
version = "0.1.0"
description = "Test task for sibdev company."
authors = ["UncleVasya <unclevasyaai@gmail.com>"]
readme = "README.md"
packages = [{include = "sibdev_job"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.3"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
django-redis = "^5.3.0"
whitenoise = "^6.5.0"
faker = "^18.11.2"
coverage = "^7.2.7"
fakeredis = "^2.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
